Add files via upload
To import data into a TensorFlow dataset object, we will use the `image_dataset_from_directory` API to load all images. The `directory` parameter is set to 'PlantVillage', which is the path to the folder containing your images. The `seed` parameter is set to 123 to ensure reproducibility by maintaining the same shuffle order each time. The `shuffle` parameter is set to True, which shuffles the images every time the dataset is accessed, helping the model learn better by preventing it from memorizing the order of the images. The `image_size` parameter resizes all images to the specified dimensions, defined by the `IMAGE_SIZE` variable. Finally, the `batch_size` parameter, defined by the `BATCH_SIZE` variable, sets the number of images to process at once, which helps manage memory and speeds up training.